requirements.txt
requests>=2.28.0
gspread>=5.10.0
google-auth>=2.22.0
